Move color chooser into plot controls
