+++ src/ert/gui/plotting/widgets/plot_controls/color_chooser.py
